extract validation code for integrated corpus
